simplified alloff function for lights

+++ VanControlCenter/LightInterface.h
@@ -54,7 +54,7 @@ public:
 
 private:
 
-    double mapf(double x, double in_min, double in_max, double out_min, double out_max);
+    uint8_t mapf(double x, double in_min, double in_max, double out_min, double out_max);
 
     light roomLight_;
     light awningLight_;
